+++ psPAS/Functions/Authentication/New-PASSession.ps1
@@ -28,13 +28,25 @@ function New-PASSession {
 			Mandatory = $true,
 			ValueFromPipeline = $true,
 			ValueFromPipelinebyPropertyName = $true,
-			ParameterSetName = 'SharedServices-Subdomain'
+			ParameterSetName = 'ISPSS-Subdomain-IdentityUser'
 		)]
 		[Parameter(
 			Mandatory = $true,
 			ValueFromPipeline = $true,
 			ValueFromPipelinebyPropertyName = $true,
-			ParameterSetName = 'SharedServices-URL'
+			ParameterSetName = 'ISPSS-URL-IdentityUser'
+		)]
+		[Parameter(
+			Mandatory = $true,
+			ValueFromPipeline = $true,
+			ValueFromPipelinebyPropertyName = $true,
+			ParameterSetName = 'ISPSS-Subdomain-ServiceUser'
+		)]
+		[Parameter(
+			Mandatory = $true,
+			ValueFromPipeline = $true,
+			ValueFromPipelinebyPropertyName = $true,
+			ParameterSetName = 'ISPSS-URL-ServiceUser'
 		)]
 		[ValidateNotNullOrEmpty()]
 		[PSCredential]$Credential,
@@ -43,7 +55,13 @@ function New-PASSession {
 			Mandatory = $true,
 			ValueFromPipeline = $false,
 			ValueFromPipelinebyPropertyName = $true,
-			ParameterSetName = 'SharedServices-Subdomain'
+			ParameterSetName = 'ISPSS-Subdomain-IdentityUser'
+		)]
+		[Parameter(
+			Mandatory = $true,
+			ValueFromPipeline = $false,
+			ValueFromPipelinebyPropertyName = $true,
+			ParameterSetName = 'ISPSS-Subdomain-ServiceUser'
 		)]
 		[string]$TenantSubdomain,
 
@@ -98,18 +116,16 @@ function New-PASSession {
 		[string]$BaseURI,
 
 		[Parameter(
-			Mandatory = $false,
+			Mandatory = $true,
 			ValueFromPipeline = $false,
 			ValueFromPipelinebyPropertyName = $true,
-			ParameterSetName = 'SharedServices-Subdomain'
+			ParameterSetName = 'ISPSS-URL-IdentityUser'
 		)]
-		[string]$IdentitySubdomain,
-
 		[Parameter(
 			Mandatory = $true,
 			ValueFromPipeline = $false,
 			ValueFromPipelinebyPropertyName = $true,
-			ParameterSetName = 'SharedServices-URL'
+			ParameterSetName = 'ISPSS-URL-ServiceUser'
 		)]
 		[string]$IdentityTenantURL,
 
@@ -117,9 +133,43 @@ function New-PASSession {
 			Mandatory = $true,
 			ValueFromPipeline = $false,
 			ValueFromPipelinebyPropertyName = $true,
-			ParameterSetName = 'SharedServices-URL'
+			ParameterSetName = 'ISPSS-URL-IdentityUser'
+		)]
+		[Parameter(
+			Mandatory = $true,
+			ValueFromPipeline = $false,
+			ValueFromPipelinebyPropertyName = $true,
+			ParameterSetName = 'ISPSS-URL-ServiceUser'
 		)]
 		[string]$PrivilegeCloudURL,
+
+		[Parameter(
+			Mandatory = $true,
+			ValueFromPipeline = $false,
+			ValueFromPipelinebyPropertyName = $true,
+			ParameterSetName = 'ISPSS-Subdomain-IdentityUser'
+		)]
+		[Parameter(
+			Mandatory = $true,
+			ValueFromPipeline = $false,
+			ValueFromPipelinebyPropertyName = $true,
+			ParameterSetName = 'ISPSS-URL-IdentityUser'
+		)]
+		[switch]$IdentityUser,
+
+		[Parameter(
+			Mandatory = $true,
+			ValueFromPipeline = $false,
+			ValueFromPipelinebyPropertyName = $true,
+			ParameterSetName = 'ISPSS-Subdomain-ServiceUser'
+		)]
+		[Parameter(
+			Mandatory = $true,
+			ValueFromPipeline = $false,
+			ValueFromPipelinebyPropertyName = $true,
+			ParameterSetName = 'ISPSS-URL-ServiceUser'
+		)]
+		[switch]$ServiceUser,
 
 		[parameter(
 			Mandatory = $true,
@@ -398,7 +448,7 @@ function New-PASSession {
 
 		Switch ($PSCmdlet.ParameterSetName) {
 
-			'SharedServices-URL' {
+			( { $PSItem -match '^ISPSS-URL' } ) {
 
 				#Ensure URLs are in expected format
 				#Remove trailing space and PasswordVault (if provided in PrivilegeCloudURL)
@@ -406,45 +456,27 @@ function New-PASSession {
 				$PrivilegeCloudURL = $PrivilegeCloudURL -replace '/$', ''
 				$PrivilegeCloudURL = $PrivilegeCloudURL -replace '/PasswordVault$', ''
 
-				#Set Required URL Values for Request
-				$LogonRequest['Uri'] = "$IdentityTenantURL/oauth2/platformtoken"
-				$Uri = "$PrivilegeCloudURL/$PVWAAppName"
+			}
+
+			( { $PSItem -match '^ISPSS-SubDomain' } ) {
+
+				$SharedServicesURLs = Find-SharedServicesURL -subdomain $TenantSubdomain
+
+				$IdentityTenantURL = $SharedServicesURLs | Select-Object -ExpandProperty identity_user_portal | Select-Object -ExpandProperty api
+				$PrivilegeCloudURL = $SharedServicesURLs | Select-Object -ExpandProperty pcloud | Select-Object -ExpandProperty api
 
 			}
 
-			'SharedServices-Subdomain' {
+			( { $PSItem -match '^ISPSS-.*-.*User$' } ) {
 
-				#Most Shared Services subdomains for Identity & Privilege Cloud tenants will be identical
-				If ($PSBoundParameters.Keys -notcontains 'IdentitySubdomain') {
-					$IDSubdomain = $TenantSubdomain
-				} Else {
-					#If different, use specified subdomain for Identity
-					$IDSubdomain = $IdentitySubdomain
-				}
+				#IdentityUser/ServiceUser LogonRequest for New-IDSession/New-IDPlatformToken
+				$LogonRequest['Uri'] = $IdentityTenantURL
+				$LogonRequest['Credential'] = $Credential
 
-				$LogonRequest['Uri'] = "https://${IDSubdomain}.id.cyberark.cloud/oauth2/platformtoken"  #hardcode Shared Services auth
+				#URL for P Cloud API Operations
+				$Uri = "${PrivilegeCloudURL}/$PVWAAppName"
 
-				#Build URL
-				$Uri = "https://${TenantSubdomain}.privilegecloud.cyberark.cloud/$PVWAAppName"
-
-			}
-
-			( { $PSItem -match '^SharedServices-' } ) {
-
-				$Body = @{
-
-					grant_type    = 'client_credentials'
-					#Add user name from credential object
-					client_id     = $($Credential.UserName)
-					#Add decoded password value from credential object
-					client_secret = $($Credential.GetNetworkCredential().Password)
-
-				}
-
-				$LogonRequest['Body'] = $Body
-				$LogonRequest['ContentType'] = 'application/x-www-form-urlencoded'
 				break
-
 			}
 
 			'integrated' {
@@ -597,8 +629,30 @@ function New-PASSession {
 
 			try {
 
-				#Send Logon Request
-				$PASSession = Invoke-PASRestMethod @LogonRequest
+				switch ($PSCmdlet.ParameterSetName) {
+					( { $PSItem -match '^ISPSS' } ) {
+						#Check IdentityCommand module available
+						if (-not (Get-Module IdentityCommand)) {
+							try { Import-Module IdentityCommand -ErrorAction Stop }
+							catch { throw 'Failed to import IdentityCommand: Install the IdentityCommand Module and try again.' }
+						}
+					}
+					( { $PSItem -match '^ISPSS-.*-IdentityUser$' } ) {
+						#Perform Identity User Authentication using IdentityCommand module
+						$PASSession = New-IDSession -tenant_url $LogonRequest['Uri'] -Credential $LogonRequest['Credential']
+						break
+					}
+					( { $PSItem -match '^ISPSS-.*-ServiceUser$' } ) {
+						#Perform Identity User Authentication using IdentityCommand module
+						$PASSession = New-IDPlatformToken -tenant_url $LogonRequest['Uri'] -Credential $LogonRequest['Credential']
+						break
+					}
+					default {
+						#Send Logon Request
+						$PASSession = Invoke-PASRestMethod @LogonRequest
+						break
+					}
+				}
 
 				If ($null -ne $PASSession.UserName) {
 
@@ -685,6 +739,19 @@ function New-PASSession {
 
 							#Shared Service access_token.
 							$CyberArkLogonResult = "$($PASSession.token_type) $($PASSession.access_token)"
+
+							#Make the IdentityCommand WebSession available in the psPAS module scope
+							Set-Variable -Name WebSession -Value $($PSItem.GetWebSession()) -Scope Script
+
+						}
+
+						( { $null -ne $PSItem.Token } ) {
+
+							#Shared Services Identity User Bearer Token
+							$CyberArkLogonResult = "Bearer $($PASSession.Token)"
+
+							#Make the IdentityCommand WebSession available in the psPAS module scope
+							Set-Variable -Name WebSession -Value $($PSItem.GetWebSession()) -Scope Script
 
 						}